requirements.txt
requests
urllib3
colorama
discord.py
python-dotenv
aiohttp
pickle
psutil
tzlocal
json
customtkinter
CTkMessagebox
pywinstyless